discoveredLabs/fastapi-app/requirements copy.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.0
httpx==0.27.0
python-dotenv==1.0.0
pydantic-settings==2.5.0
google-api-python-client==2.149.0
aiohttp==3.9.1

